- MAJROR Refractoring: switched to effects and visual passes instead of scene passes - added VisualEffect and all the needed clases inheriting from it - added Lit/Unlit effects Unlit2DTextureArrayVisualEffect, Unlit2DTextureVisualEffect, UnlitColorAttributeVisualEffect, UnlitColorVisualEffect, UnlitCubemapTextureVisualEffect Lit2DTextureVisualEffect, LitColorAttributeVisualEffect, LitColorVisualEffect, LitEnvironmentMappingTextureVisualEffect, LitMirrorTextureVisualEffect, LitShadowColorAttributeVisualEffect - added VisualPass class and its VulkanVisualPass counterpart (Vulkan backend) - moved all Vulkan Resources like: descriptor sets, layouts, pipeline from VulkanRenderer to VulkanVisualPass - renderpass, framebufers are now part of the VulkanRenderer logic instead of the VisualPass ones - added the required shaders for all supported effects (both lit and unlit) - added support for projected shadows - added support for point lights (needed for shadows implementation) - improved light support (light colour moved to base class) - improved support for offscreen rendering techniques (mirror and shadows) - removed ScenePass logic and related classes - changed camera pos, light dir/pos/colour from vec3 to vec4 in shaders to align properly with teh std140 glsl memory alignment - added explicit support for different uniform types and uniform vars in shaders (see GLSLShaderTypes.h) - improvements and cleanup in GLSLParser class - improved gltf2Loader class with better LoadingFlags - camera fov is now a float instead of int - VisualComponent is now just a wrapper class, the main functionality has been moved to the VisualPass class - PrimitiveTopology, FaceWinding and PolygonMode are no longer defined in GeometricPrimitive, but in VisualPass class - updated VulkanRenderer with all the new logic, some improvemnts to draw, cleanup - pipeline queries are now done per VisualPass in VulkanRenderer - updated VulkanCommon header (depth defines) - added Bind() to VulkanGraphicsPipeline, VulkanCOmputePipeline and VulkanDescriptorSet classes - improved support for depth textures/render targets in VulkanTexture - moved VertexInputRate adn VertexInputBinding from VertexBuffer, VulkanVertexBuffer to VertexFormat, VulkanVertexFormat - improved support for sampling and texture usage in Texture and RenderTarget classes - improved uniform support in UniformBuffer - cleanup in RenderQueue class - cleanup in Renderer class - removed unused classes: RenderFramebuffer, VulkanRenderFramebuffer, VulkanRenderTarget - improved allowed passess in Node and GeometryNode classes - added names to all nodes in all samples - aligned all sample code with the new logic - removed shaders and other unneeded resources from each sample - added Shadows sample - small improvements - added a screenshots per sample app

+++ LibGraphicsEngine/src/Graphics/Cameras/Camera.cpp
@@ -50,14 +50,14 @@ void Camera::UpdateViewMatrix()
 
 void Camera::UpdatePerspectiveProjectionMatrix()
 {
-	assert(mFOVy > 0);
-	assert(mAspectRatio > 0);
-	assert(mZNear > 0);
-	assert(mZFar > 0);
+	assert(mFOVy > 0.0f);
+	assert(mAspectRatio > 0.0f);
+	assert(mZNear > 0.0f);
+	assert(mZFar > 0.0f);
 
 	float correctionFactor = ComputePerspectiveProjectionCorrectionFactor();
 
-	mProjection = glm::perspective(glm::radians(static_cast<float32_t>(mFOVy)), mAspectRatio,
+	mProjection = glm::perspective(glm::radians(mFOVy), mAspectRatio,
 		mZNear * correctionFactor, mZFar * correctionFactor);
 
 	//#if defined(RIGHT_HAND_COORDINATES)
@@ -104,12 +104,12 @@ float Camera::ComputePerspectiveProjectionCorrectionFactor() const
 	return cameraAltitude;
 }
 
-void Camera::UpdatePerspectiveProjectionMatrix(int32_t fovy, float32_t aspect, float32_t zNear, float32_t zFar)
+void Camera::UpdatePerspectiveProjectionMatrix(float32_t fovy, float32_t aspect, float32_t zNear, float32_t zFar)
 {
-	assert(fovy > 0);
-	assert(aspect > 0);
-	assert(zNear > 0);
-	assert(zFar > 0);
+	assert(fovy > 0.0f);
+	assert(aspect > 0.0f);
+	assert(zNear > 0.0f);
+	assert(zFar > 0.0f);
 
 	mFOVy = fovy;
 	mAspectRatio = aspect;
@@ -121,8 +121,8 @@ void Camera::UpdatePerspectiveProjectionMatrix(int32_t fovy, float32_t aspect, f
 
 void Camera::UpdateOrthographicProjectionMatrix(float32_t left, float32_t right, float32_t bottom, float32_t top, float32_t zNear, float32_t zFar)
 {
-	assert(zNear > 0);
-	assert(zFar > 0);
+	assert(zNear > 0.0f);
+	assert(zFar > 0.0f);
 
 	// NOTE! No FOV for the ortho projection
 	LOG_INFO("No FOV is used for the ortho projection! It remains unchanged!");
@@ -174,7 +174,7 @@ const glm::vec3& Camera::GetUp() const
 	return mUp;
 }
 
-int Camera::GetFOV() const
+float32_t Camera::GetFOV() const
 {
 	return mFOVy;
 }
@@ -234,7 +234,7 @@ void Camera::SetUp(const glm::vec3& up)
 	mUp = up;
 }
 
-void Camera::SetFOV(int32_t fov)
+void Camera::SetFOV(float32_t fov)
 {
 	mFOVy = fov;
 }
